split kernel into kernel extensions for relevant pieces
